+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove `vscode-graphiql-explorer` integration.
 - Add option to add query boilerplate for preloaded query.
 - Add step for selecting what type to expand when generating a query that uses the node interface/node top level field.
+- Add step for autogenerating any needed variables etc when codegenning mutations.
 
 ## 0.3.6